Updated the sense lib with the new DSE and removed unused and deprecated stuff
