+++ sense-android-library/src/nl/sense_os/service/external_sensors/ZephyrHxmRegistrator.java
@@ -1,6 +1,5 @@
 package nl.sense_os.service.external_sensors;
 
-import nl.sense_os.service.commonsense.SensorRegistrator;
 import nl.sense_os.service.constants.SenseDataTypes;
 import nl.sense_os.service.constants.SensorData.SensorNames;
 import android.content.Context;
@@ -17,13 +16,12 @@ import android.content.Context;
  * 
  * @author Steven Mulder <steven@sense-os.nl>
  */
-public class ZephyrHxmRegistrator extends SensorRegistrator {
+public class ZephyrHxmRegistrator{
 
-    public ZephyrHxmRegistrator(Context context) {
-        super(context);
+    public ZephyrHxmRegistrator() {
+
     }
 
-    @Override
     public boolean verifySensorIds(String deviceType, String deviceUuid) {
         // preallocate
         String name, displayName, description, dataType, value;
@@ -35,8 +33,8 @@ public class ZephyrHxmRegistrator extends SensorRegistrator {
         description = "HxM " + deviceType;
         dataType = SenseDataTypes.INT;
         value = "0";
-        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
-                deviceUuid);
+//        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
+//                deviceUuid);
 
         // match speed sensor
         name = SensorNames.SPEED;
@@ -44,8 +42,8 @@ public class ZephyrHxmRegistrator extends SensorRegistrator {
         description = "HxM " + deviceType;
         dataType = SenseDataTypes.FLOAT;
         value = "0.0";
-        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
-                deviceUuid);
+//        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
+//                deviceUuid);
 
         // match distance sensor
         name = SensorNames.DISTANCE;
@@ -53,8 +51,8 @@ public class ZephyrHxmRegistrator extends SensorRegistrator {
         description = "HxM " + deviceType;
         dataType = SenseDataTypes.FLOAT;
         value = "0.0";
-        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
-                deviceUuid);
+//        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
+//                deviceUuid);
 
         // match battery charge sensor
         name = SensorNames.BATTERY_CHARGE;
@@ -62,8 +60,8 @@ public class ZephyrHxmRegistrator extends SensorRegistrator {
         description = "HxM " + deviceType;
         dataType = SenseDataTypes.INT;
         value = "0";
-        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
-                deviceUuid);
+//        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
+//                deviceUuid);
 
         // match strides sensor
         name = SensorNames.STRIDES;
@@ -71,8 +69,8 @@ public class ZephyrHxmRegistrator extends SensorRegistrator {
         description = "HxM " + deviceType;
         dataType = SenseDataTypes.INT;
         value = "0";
-        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
-                deviceUuid);
+//        success &= checkSensor(name, displayName, dataType, description, value, deviceType,
+//                deviceUuid);
 
         return success;
     }
